+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added size classes for material icons.
 - Added hover version of item-list.
 - Added examples tab to topbar which redirects to our example pages.
+- Added dashboard theme with chart component.
 
 ### Changed